Add to GenerateOrder page

+++ AgateMountainWarehouse.Client/ViewModels/SalesOrderViewModel.cs
@@ -4,7 +4,7 @@ public class SalesOrderViewModel
 {
     public Guid Id { get; set; }
     public bool IsPaid { get; set; } = false;
-    public int TotalCost { get; set; }
+    public double TotalCost { get; set; }
     public List<SalesOrderItemViewModel> OrderItems { get; set; } = new List<SalesOrderItemViewModel>();
     public DateTime CreatedDateTime { get; set; }
     public DateTime? UpdatedDateTime { get; set; }
